requirements.txt
jinja2
telethon
apiai==1.2.3
fbmq==2.1.0
python-telegram-bot==8.1.1
python-decouple==3.1
peewee==2.10.1
logzero==1.3.1
psycopg2==2.7.3.2
peewee_migrate==0.14.0
ruamel.yaml
